ref: refactor firebase config to service folder
