+++ src/views/ChatRoomGroup.vue
@@ -104,7 +104,7 @@
 import { reactive, toRefs, onMounted, ref, computed, onBeforeMount, onUpdated } from 'vue'
 import moment from 'moment'
 import Chat from '../components/Chat.vue'
-import {auth,database, firestore} from '../firebase'
+import {auth,database, firestore} from '../service/firebase'
 import { useRoute, useRouter } from 'vue-router'
 import MenuOption from '../components/MenuOption.vue'
 import Spinner from '../components/Spinner.vue'
