+++ src/views/ProfilePage.vue
@@ -86,7 +86,7 @@
 
 <script>
 import {onBeforeMount, onMounted, reactive, toRefs } from 'vue'
-import { firestore, storage } from '../firebase'
+import { firestore, storage } from '../service/firebase'
 import { useRouter } from 'vue-router'
 import Spinner from '../components/Spinner.vue'
 import U from '../components/svg/U.vue'
